pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Group expense tracking application with FastAPI and MySQL"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.121.3",
    "uvicorn>=0.38.0",
    "jinja2>=3.1.6",
    "mysql-connector-python>=9.5.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "ruff>=0.14.5",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.pyright]
typeCheckingMode = "basic"
pythonVersion = "3.14"
